+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [zeebe-bpmn-moddle](https://github.com/camunda/zeebe-bpmn
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: add `zeebe:agentDefinition` extension element ([#94](https://github.com/camunda/zeebe-bpmn-moddle/pull/94))
+
 ## 1.17.0
 
 * `FEAT`: add `businessId` to `zeebe:CalledElement` for Call Activity Business ID propagation ([#90](https://github.com/camunda/zeebe-bpmn-moddle/pull/90))